split out configure functions into separate file
